chore: add DrupalAttribute toString test
Add jest to test DrupalAttribute's `toString` method. For this, `DrupalAttribute` was moved into its own file so it can be imported in the test without adding it to the export of index.js.
