+++ package.json
@@ -12,6 +12,7 @@
     "url": "git@github.com:miyagi-dev/twig-drupal.git"
   },
   "scripts": {
+    "test": "jest",
     "release": "standard-version"
   },
   "dependencies": {
@@ -20,5 +21,9 @@
   },
   "peerDependencies": {
     "@miyagi/core": "^3.1.7"
+  },
+  "devDependencies": {
+    "@jest/globals": "^28.1.3",
+    "jest": "^28.1.3"
   }
 }
